Add files via upload
The task:
On the Prozorro analytics portal (https://bi.prozorro.org/sense/app/fba3f2f2-cf55-40a0-a79f-b74f5ce947c2/overview) find data on the tenders held for 2020, which have the status of completed and competitive. According to the generalized CPV category (up to two digits: for (09120000-6 Газове паливо) the generalized category will be 09), display the total budget of lots, the number of lots, and the number of tenders.